fix: Resolve issue with page responsiveness
